auth

+++ src/main/java/com/yevheniir/hwtp/controller/HwtpController.java
@@ -1,9 +1,11 @@
 package com.yevheniir.hwtp.controller;
 
+import com.yevheniir.hwtp.exception.ForbiddenException;
 import com.yevheniir.hwtp.model.Order;
 import com.yevheniir.hwtp.model.Stuff;
 import com.yevheniir.hwtp.repository.OrderRepository;
 import com.yevheniir.hwtp.repository.UserDetailsRepo;
+import com.yevheniir.hwtp.service.AuthService;
 import com.yevheniir.hwtp.service.EmailService;
 import com.yevheniir.hwtp.service.HwtpService;
 import com.yevheniir.hwtp.service.StorageService;
@@ -23,6 +25,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -44,6 +47,9 @@ public class HwtpController {
     @Autowired
     private EmailService emailService;
 
+    @Autowired
+    private AuthService authService;
+
     @GetMapping("stuff")
     List<Stuff> getStuff() {
         return hwtpService.getStuff();
@@ -51,14 +57,20 @@ public class HwtpController {
 
 
     @PostMapping("stuff")
-    Stuff addStuff(@RequestBody Stuff stuff) {
-        return hwtpService.addStuff(stuff);
+    Stuff addStuff(@RequestBody Stuff stuff, @RequestHeader(value="Auth") String auth ) {
+        if (authService.validToken(auth)) {
+            return hwtpService.addStuff(stuff);
+        }
+        throw new ForbiddenException();
     }
 
-
     @DeleteMapping("stuff/{id}")
-    public void deleteStuff(@PathVariable Long id) {
-        hwtpService.deleteStuff(id);
+    public void deleteStuff(@PathVariable Long id, @RequestHeader(value="Auth") String auth ) {
+        if (authService.validToken(auth)) {
+            hwtpService.deleteStuff(id);
+        } else{
+            throw new ForbiddenException();
+        }
     }
 
     @PostMapping("stuff/file")
@@ -77,18 +89,30 @@ public class HwtpController {
     }
 
     @DeleteMapping("orders/{id}")
-    void deleteOrder(@PathVariable Long id) throws MessagingException {
-        hwtpService.deleteOrder(id);
+    void deleteOrder(@PathVariable Long id, @RequestHeader(value="Auth") String auth ) throws MessagingException {
+        if (authService.validToken(auth)) {
+            hwtpService.deleteOrder(id);
+        } else{
+            throw new ForbiddenException();
+        }
     }
 
     @PostMapping("orders/{id}")
-    void commentOrder(@PathVariable Long id, @RequestBody String comment) throws MessagingException {
-        hwtpService.commentOrder(id, comment);
+    void commentOrder(@PathVariable Long id, @RequestBody String comment, @RequestHeader(value="Auth") String auth ) throws MessagingException {
+        if (authService.validToken(auth)) {
+            hwtpService.commentOrder(id, comment);
+        } else{
+            throw new ForbiddenException();
+        }
     }
 
     @DeleteMapping("orders/cancel/{id}")
-    void cancelOrder(@PathVariable Long id) throws MessagingException {
-        hwtpService.canceOlrder(id);
+    void cancelOrder(@PathVariable Long id, @RequestHeader(value="Auth") String auth ) throws MessagingException {
+        if (authService.validToken(auth)) {
+            hwtpService.canceOlrder(id);
+        } else{
+            throw new ForbiddenException();
+        }
     }
 
     @PostMapping("orders")
@@ -118,14 +142,31 @@ public class HwtpController {
 //
 //        return "index";
 //    }
-//
+
 //    @GetMapping("users")
 //    List<User> getusers() {
 //        return userDetailsRepo.findAll();
+//    }
+//
+//    @RequestMapping(value = "/controller", method = RequestMethod.GET)
+//    @ResponseBody
+//    public ResponseEntity sendViaResponseEntity() {
+//        return new ResponseEntity(HttpStatus.NOT_ACCEPTABLE);
 //    }
 //
 //    @GetMapping("logout")
 //    void userLogout() {
 //        return;
 //    }
+
+    @PostMapping("password")
+    public List<String> validatePassword(@RequestBody String password) {
+        if (password.equals("password")) {
+            List<String> l = new ArrayList<>();
+            l.add(this.authService.createToken());
+            return l ;
+        } else {
+            return null;
+        }
+    }
 }
